Pit summary in team summary view is now defined with an XML layout like the rest of the app. Minor refactoring.

+++ android/src/main/java/org/wildstang/wildrank/android/customviews/SerializableNumberView.java
@@ -9,83 +9,83 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import org.wildstang.wildrank.R;
+import org.wildstang.wildrank.android.R;
 
 import java.util.Locale;
 
 public class SerializableNumberView extends JSONSerializableView {
 
-	private TextView labelView;
-	private EditText valueView;
+    private TextView labelView;
+    private EditText valueView;
 
-	public SerializableNumberView(Context context, AttributeSet attrs) {
-		super(context, attrs);
+    public SerializableNumberView(Context context, AttributeSet attrs) {
+        super(context, attrs);
 
-		LayoutInflater inflater = LayoutInflater.from(context);
-		inflater.inflate(R.layout.custom_view_number, this, true);
+        LayoutInflater inflater = LayoutInflater.from(context);
+        inflater.inflate(R.layout.custom_view_number, this, true);
 
-		TypedArray a = context.obtainStyledAttributes(attrs, R.styleable.JSONSerializableView, 0, 0);
-		String label = a.getString(R.styleable.JSONSerializableView_label);
-		a.recycle();
+        TypedArray a = context.obtainStyledAttributes(attrs, R.styleable.JSONSerializableView, 0, 0);
+        String label = a.getString(R.styleable.JSONSerializableView_label);
+        a.recycle();
 
-		labelView = (TextView) findViewById(R.id.label);
-		labelView.setText(label);
+        labelView = (TextView) findViewById(R.id.label);
+        labelView.setText(label);
 
-		valueView = (EditText) findViewById(R.id.value);
-		valueView.setClickable(false);
-		// When we receive focus after pressing "Next", pass focus onto the edittext
-		this.setFocusable(true);
-		this.setFocusableInTouchMode(true);
-		this.setOnFocusChangeListener(new OnFocusChangeListener() {
-			@Override
-			public void onFocusChange(View v, boolean hasFocus) {
-				if (hasFocus) {
-					valueView.requestFocus();
-				}
+        valueView = (EditText) findViewById(R.id.value);
+        valueView.setClickable(false);
+        // When we receive focus after pressing "Next", pass focus onto the edittext
+        this.setFocusable(true);
+        this.setFocusableInTouchMode(true);
+        this.setOnFocusChangeListener(new OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View v, boolean hasFocus) {
+                if (hasFocus) {
+                    valueView.requestFocus();
+                }
 
-			}
-		});
-	}
+            }
+        });
+    }
 
-	public void setValue(double value) {
-		valueView.setText(formatDouble(value));
-	}
+    public void setValue(double value) {
+        valueView.setText(formatDouble(value));
+    }
 
-	@Override
-	public void writeContentsToBundle(Bundle b) {
-		if (valueView.getText().toString().trim().length() != 0) {
-			// Only save the value if it isn't empty
-			b.putDouble(key, Double.parseDouble(valueView.getText().toString()));
-		} else {
-			b.putDouble(key, Double.valueOf(0.0));
-		}
-	}
+    @Override
+    public void writeContentsToBundle(Bundle b) {
+        if (valueView.getText().toString().trim().length() != 0) {
+            // Only save the value if it isn't empty
+            b.putDouble(key, Double.parseDouble(valueView.getText().toString()));
+        } else {
+            b.putDouble(key, 0.0);
+        }
+    }
 
-	public static String formatDouble(double d) {
-		if (d == (int) d) {
-			return String.format(Locale.US, "%d", (int) d);
-		} else {
-			return String.format(Locale.US, "%s", d);
-		}
-	}
+    public static String formatDouble(double d) {
+        if (d == (int) d) {
+            return String.format(Locale.US, "%d", (int) d);
+        } else {
+            return String.format(Locale.US, "%s", d);
+        }
+    }
 
-	@Override
-	public void restoreFromJSONObject(Object object) {
-		if (object instanceof Integer) {
-			setValue((double) ((Integer) object).intValue());
-		} else if (object instanceof Double) {
-			setValue((Double) object);
-		}
-	}
+    @Override
+    public void restoreFromJSONObject(Object object) {
+        if (object instanceof Integer) {
+            setValue((double) (Integer) object);
+        } else if (object instanceof Double) {
+            setValue((Double) object);
+        }
+    }
 
-	@Override
-	public boolean isComplete(boolean highlight) {
-		if (valueView.getText().toString().isEmpty()) {
-			isComplete = false;
-		} else {
-			isComplete = true;
-		}
-		return super.isComplete(highlight);
-	}
+    @Override
+    public boolean isComplete(boolean highlight) {
+        if (valueView.getText().toString().isEmpty()) {
+            isComplete = false;
+        } else {
+            isComplete = true;
+        }
+        return super.isComplete(highlight);
+    }
 
 }
